pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecmprep"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anyio==4.11.0",
    "attrs==25.3.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "lxml==6.0.2",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pip==25.2",
    "pluggy==1.6.0",
    "pygments==2.19.2",
    "pytest==8.4.2",
    "pytest-asyncio==0.23.7",
    "pytest-trio==0.8.0",
    "python-docx==1.2.0",
    "Pillow==11.0.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "trio==0.31.0",
    "typing-extensions==4.15.0"
]

[project.scripts]
ecmprep = "mypkg.cli.main:entry_point"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["**/*"]  # 필요 시

[tool.pytest.ini_options]
markers = [
    "asyncio: mark a test as async"
]
